requirements.txt
pandas~=2.2.1
matplotlib~=3.8.3
requests==2.26.0
gspread==4.0.1
urllib3==1.26.7
oauth2client==4.1.3
datetime==4.3
pywin32~=306
pynput==1.7.3
Pillow~=10.2.0
logging~=0.4.9.6
google-api-python-client~=2.120.0
setuptools~=69.1.1